Reduce fixed sessions graph view to last 24 hrs
For now we simply limit the data only during the loading graph
view again. In the future it could be improved to keep it more
consistent (figure out what with mobile sessions; live update
the graph view - not only during the loading; or to not display
fixed sessions measusrements in the phone at all).